Log Detailed Version
GraalVM has two different versions, the GraalVM version, and the Java version
that it's implementing.  Previously, the various buildpacks only printed the
Java version directly, although a close reading of the logs would give you the
other version.  This extra work wasn't the best UX, so this change adds a log
line, before the native-image build begins, with the detailed version
information about the version of GraalVM.

[resolves #1]

Signed-off-by: Ben Hale <bhale@vmware.com>