Compatibility with Spring Boot 2.4.2
In Boot 2.4.2 the format of classpath.idx changed to include the relative path of each jar that should be added to the classpath. Previous in only provided a list of jar names. When generating the classpath, the buildpack will now only prepend the value of Spring-Boot-Lib from the manifest to the indexed entries if the entries are naked file names. It will use the exact path provided if the entry contains a '/'

Also, adds all native-image arguments to the layer metadata (not just additional user provided arguments) so that changes in buildpack computed arguments trigger a rebuild.

Signed-off-by: Emily Casey <ecasey@vmware.com>