Do not require Start-Class in MANIFEST.MF
Prior to this PR, it was required that you have a Start-Class entry in MANIFEST.MF. That is a Spring Boot convention though and not everyone has that entry.

To support non-Spring Boot apps better, this PR relaxes the requirement. If present, the Start-Class will be used. If not present, it will fall back to Main-Class which is a standard Java executable JAR convention. If neither are present, an error will occur.

Resolves #83.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>