Support building from JAR files
Prior to this PR, native-image builds would happen with the assumption that the contents of the `/workspace` directory was an exploded JAR file. The buildpack would configure options to build from that location. Recently, buildpacks have gained the capability to build and persist multiple files. This results in the `/workspace` directory not meeting the previous expectation on structure.

This PR continues to support building from an exploded JAR file, but also supports building from a JAR file. This requires the user to specify a pattern (Glob match) or file name which points to the JAR file that should the target for the native-image build. When this is set, the buildpack will adjust the build arguments to build from the JAR file.

The `$BP_NATIVE_IMAGE_BUILT_ARTIFACT` argument is what controls this behavior. If the `/workspace` directory does not contain an exploded JAR structure, it must have `META-INF/MANIFEST.MF`, then we'll look for files that match the pattern set in `$BP_NATIVE_IMAGE_BUILT_ARTIFACT` and attempt to build from the JAR file at that location.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>