Also copy .so files when doing a native-image build
Some applications, when native-image runs, will also generate shared libraries which are required to run the application. These libraries get put next to the generated binary file. This change modifies the buildpack to copy these shared libraries after it copies the binary to the application directory.

In a nutshell, native-image runs, writes binary and shared libs to a cached layer. We then copy files from the cached layer to the application directory (i.e. /workspace) which results in them being available in the application image.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>